+++ misc/metrics/CHANGELOG.md
@@ -1,4 +1,22 @@
-# 0.7.0 [unreleased]
+# 0.8.0 [unreleased]
+
+- Update to `libp2p-swarm` `v0.38.0`.
+
+- Update to `libp2p-dcutr` `v0.5.0`.
+
+- Update to `libp2p-ping` `v0.38.0`.
+
+- Update to `libp2p-identify` `v0.38.0`.
+
+- Update to `libp2p-relay` `v0.11.0`.
+
+- Update to `libp2p-kad` `v0.39.0`.
+
+- Track number of connected nodes supporting a specific protocol via the identify protocol. See [PR 2734].
+
+[PR 2734]: https://github.com/libp2p/rust-libp2p/pull/2734/
+
+# 0.7.0
 
 - Update to `libp2p-core` `v0.34.0`.