+++ protocols/autonat/CHANGELOG.md
@@ -1,4 +1,10 @@
-# 0.5.0 [unreleased]
+# 0.6.0 [unreleased]
+
+- Update to `libp2p-swarm` `v0.38.0`.
+
+- Update to `libp2p-request-response` `v0.20.0`.
+
+# 0.5.0
 
 - Update to `libp2p-core` `v0.34.0`.