+++ protocols/ping/CHANGELOG.md
@@ -1,4 +1,12 @@
-# 0.37.0 [unreleased]
+# 0.38.0 [unreleased]
+
+- Update to `libp2p-swarm` `v0.38.0`.
+
+- Expose `PROTOCOL_NAME`. See [PR 2734].
+
+[PR 2734]: https://github.com/libp2p/rust-libp2p/pull/2734/
+
+# 0.37.0
 
 - Update to `libp2p-core` `v0.34.0`.